Completed Love scoring
Used a list of love-related keywords to score songs related to love. love score now is not counting the UNIQUE love words, but ABSOLUTE count of love-related words per song. It should be reasonable to assume that the importance of 'love' increases with the repeat of words it is the more they will repeat the words.